refactor: keep bench comparison slices audit-clean
